refactor: extract GlossaryPanel from SeriesDetail.tsx
